done proto slider
