+++ src/js/main.js
@@ -1,8 +1,11 @@
-import Slider from './modules/slider';
+import MainSlider from './modules/slider/slider-main';
 import VideoPlayer from './modules/playVideo';
 
-window.addEventListener('DOMContentLoaded', () =>{
-	const slider = new Slider('.page', '.next');
+window.addEventListener('DOMContentLoaded', () => {
+	const slider = new MainSlider({
+		slider: '.page',
+		buttons: '.next',
+	});
 	const videoPlayer = new VideoPlayer('.showup .play', '.overlay');
 
 	slider.render();
